toolkit/devtools/tests/mochitest/test_invisible_loader.html

<!DOCTYPE html>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript"
            src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css"
          href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript;version=1.8">
      const { utils: Cu } = Components;

      /**
       * Ensure that sandboxes created via the Dev Tools loader respect the
       * invisibleToDebugger flag.
       */
      const { DevToolsLoader } =
        Cu.import("resource://gre/modules/devtools/Loader.jsm", {});
      Cu.import("resource://gre/modules/jsdebugger.jsm");
      addDebuggerToGlobal(this);

      const COLOR_URI = "resource://gre/modules/devtools/css-color.js";

      visible_loader();
      invisible_loader();

      function visible_loader() {
        let loader = new DevToolsLoader();
        loader.invisibleToDebugger = false;
        loader.require("devtools/css-color");

        let dbg = new Debugger();
        let sandbox = loader._provider.loader.sandboxes[COLOR_URI];

        try {
          dbg.addDebuggee(sandbox);
          ok(true, "debugger added visible value");
        } catch(e) {
          ok(false, "debugger could not add visible value");
        }
      }

      function invisible_loader() {
        let loader = new DevToolsLoader();
        loader.invisibleToDebugger = true;
        loader.require("devtools/css-color");

        let dbg = new Debugger();
        let sandbox = loader._provider.loader.sandboxes[COLOR_URI];

        try {
          dbg.addDebuggee(sandbox);
          ok(false, "debugger added invisible value");
        } catch(e) {
          ok(true, "debugger could not add invisible value");
        }
      }
    </script>
  </body>
</html>
